Fix two bugs with threading that were probably causing chaos.
In the constructor of WorkerThread id(id) is of course making id
depend on uninitalised memory, we wanted id(_id). and later on we were
setting i  to 0 rather than id, so all the threads did the same
work. Hilariously, this meant we never used the bad value of id
so it didn't cause a problem.